Split up ContractsBCL into separate classes.
