+++ src/System.Private.CoreLib/src/System/Runtime/CompilerServices/ContractHelper.cs
@@ -18,182 +18,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Contracts;
 using System.Reflection;
 
-namespace System.Diagnostics.Contracts
-{
-    public static partial class Contract
-    {
-        #region Private Methods
-
-        /// <summary>
-        /// This method is used internally to trigger a failure indicating to the "programmer" that he is using the interface incorrectly.
-        /// It is NEVER used to indicate failure of actual contracts at runtime.
-        /// </summary>
-        static void AssertMustUseRewriter(ContractFailureKind kind, string contractKind)
-        {
-            // For better diagnostics, report which assembly is at fault.  Walk up stack and
-            // find the first non-mscorlib assembly.
-            Assembly thisAssembly = typeof(Contract).Assembly;  // In case we refactor mscorlib, use Contract class instead of Object.
-            StackTrace stack = new StackTrace();
-            Assembly probablyNotRewritten = null;
-            for (int i = 0; i < stack.FrameCount; i++)
-            {
-                Assembly caller = stack.GetFrame(i).GetMethod().DeclaringType.Assembly;
-                if (caller != thisAssembly)
-                {
-                    probablyNotRewritten = caller;
-                    break;
-                }
-            }
-
-            if (probablyNotRewritten == null)
-                probablyNotRewritten = thisAssembly;
-            string simpleName = probablyNotRewritten.GetName().Name;
-            System.Runtime.CompilerServices.ContractHelper.TriggerFailure(kind, SR.Format(SR.MustUseCCRewrite, contractKind, simpleName), null, null, null);
-        }
-
-        #endregion Private Methods
-
-        #region Failure Behavior
-
-        /// <summary>
-        /// Without contract rewriting, failing Assert/Assumes end up calling this method.
-        /// Code going through the contract rewriter never calls this method. Instead, the rewriter produced failures call
-        /// System.Runtime.CompilerServices.ContractHelper.RaiseContractFailedEvent, followed by 
-        /// System.Runtime.CompilerServices.ContractHelper.TriggerFailure.
-        /// </summary>
-        [System.Diagnostics.DebuggerNonUserCode]
-        static void ReportFailure(ContractFailureKind failureKind, string userMessage, string conditionText, Exception innerException)
-        {
-            if (failureKind < ContractFailureKind.Precondition || failureKind > ContractFailureKind.Assume)
-                throw new ArgumentException(SR.Format(SR.Arg_EnumIllegalVal, failureKind), nameof(failureKind));
-
-            // displayMessage == null means: yes we handled it. Otherwise it is the localized failure message
-            var displayMessage = System.Runtime.CompilerServices.ContractHelper.RaiseContractFailedEvent(failureKind, userMessage, conditionText, innerException);
-
-            if (displayMessage == null) return;
-
-            System.Runtime.CompilerServices.ContractHelper.TriggerFailure(failureKind, displayMessage, userMessage, conditionText, innerException);
-        }
-
-        /// <summary>
-        /// Allows a managed application environment such as an interactive interpreter (IronPython)
-        /// to be notified of contract failures and 
-        /// potentially "handle" them, either by throwing a particular exception type, etc.  If any of the
-        /// event handlers sets the Cancel flag in the ContractFailedEventArgs, then the Contract class will
-        /// not pop up an assert dialog box or trigger escalation policy.  Hooking this event requires 
-        /// full trust, because it will inform you of bugs in the appdomain and because the event handler
-        /// could allow you to continue execution.
-        /// </summary>
-        public static event EventHandler<ContractFailedEventArgs> ContractFailed
-        {
-            add
-            {
-                System.Runtime.CompilerServices.ContractHelper.InternalContractFailed += value;
-            }
-            remove
-            {
-                System.Runtime.CompilerServices.ContractHelper.InternalContractFailed -= value;
-            }
-        }
-        #endregion FailureBehavior
-    }
-
-    public sealed class ContractFailedEventArgs : EventArgs
-    {
-        private ContractFailureKind _failureKind;
-        private string _message;
-        private string _condition;
-        private Exception _originalException;
-        private bool _handled;
-        private bool _unwind;
-
-        internal Exception thrownDuringHandler;
-
-        public ContractFailedEventArgs(ContractFailureKind failureKind, string message, string condition, Exception originalException)
-        {
-            Debug.Assert(originalException == null || failureKind == ContractFailureKind.PostconditionOnException);
-            _failureKind = failureKind;
-            _message = message;
-            _condition = condition;
-            _originalException = originalException;
-        }
-
-        public string Message { get { return _message; } }
-        public string Condition { get { return _condition; } }
-        public ContractFailureKind FailureKind { get { return _failureKind; } }
-        public Exception OriginalException { get { return _originalException; } }
-
-        // Whether the event handler "handles" this contract failure, or to fail via escalation policy.
-        public bool Handled
-        {
-            get { return _handled; }
-        }
-
-        public void SetHandled()
-        {
-            _handled = true;
-        }
-
-        public bool Unwind
-        {
-            get { return _unwind; }
-        }
-
-        public void SetUnwind()
-        {
-            _unwind = true;
-        }
-    }
-
-    [Serializable]
-    [System.Runtime.CompilerServices.TypeForwardedFrom("mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")]
-    // Needs to be public to support binary serialization compatibility
-    public sealed class ContractException : Exception
-    {
-        private readonly ContractFailureKind _kind;
-        private readonly string _userMessage;
-        private readonly string _condition;
-
-        public ContractFailureKind Kind { get { return _kind; } }
-        public string Failure { get { return this.Message; } }
-        public string UserMessage { get { return _userMessage; } }
-        public string Condition { get { return _condition; } }
-
-        // Called by COM Interop, if we see COR_E_CODECONTRACTFAILED as an HRESULT.
-        private ContractException()
-        {
-            HResult = System.Runtime.CompilerServices.ContractHelper.COR_E_CODECONTRACTFAILED;
-        }
-
-        public ContractException(ContractFailureKind kind, string failure, string userMessage, string condition, Exception innerException)
-            : base(failure, innerException)
-        {
-            HResult = System.Runtime.CompilerServices.ContractHelper.COR_E_CODECONTRACTFAILED;
-            _kind = kind;
-            _userMessage = userMessage;
-            _condition = condition;
-        }
-
-        private ContractException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context)
-            : base(info, context)
-        {
-            _kind = (ContractFailureKind)info.GetInt32("Kind");
-            _userMessage = info.GetString("UserMessage");
-            _condition = info.GetString("Condition");
-        }
-
-
-        public override void GetObjectData(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context)
-        {
-            base.GetObjectData(info, context);
-            info.AddValue("Kind", _kind);
-            info.AddValue("UserMessage", _userMessage);
-            info.AddValue("Condition", _condition);
-        }
-    }
-}
-
-
 namespace System.Runtime.CompilerServices
 {
     public static partial class ContractHelper
